* ParseAligns/ParseAligns.cpp (stats): Should be size_t, not int.
git-svn-id: https://svn.bcgsc.ca/svn/abyss/trunk@52814 ad02020f-2a26-4133-947e-1829317c1416

+++ ParseAligns/ParseAligns.cpp
@@ -79,13 +79,13 @@ static const struct option longopts[] = {
 };
 
 static struct {
-	int alignments;
-	int numDifferent;
-	int numSame;
-	int numMisoriented;
-	int numMissed;
-	int numMulti;
-	int numSplit;
+	size_t alignments;
+	size_t numDifferent;
+	size_t numSame;
+	size_t numMisoriented;
+	size_t numMissed;
+	size_t numMulti;
+	size_t numSplit;
 } stats;
 
 static ostream& pairedAlignFile = cout;
